requirements.txt
django
djangorestframework
psycopg2
geopandas
requests
pandas
lxml